add `sqrt` module; write documentation for all `sqrt` stuff
